Move command modules to startup assembly
